# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Five 网页验证版接口
  version: 1.0.0
  summary: 公开网页、单人后台与后端共同使用的 P0 接口契约
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    本文件是 Five P0 接口字段、状态、错误码和版本保护规则的唯一事实源。
    `contentVersion`、各种 ID 和修订号都是不透明值，调用方不得解析其字符串。
    P0 只有网页端，因此接口不包含 `platform`、`sourcePlatform` 或 `targetPlatform`。
    `channelId` 只用于区分链接分发来源，不得改变日期、档位、正文或图片。
    所有 `/admin/api/v1` 响应（包括错误）必须使用 `Cache-Control: no-store`，防止退出后从缓存读取后台内容或核对凭证。
servers:
  - url: /
    description: 与网页同源的后端
...
security: []
paths:
...
  /admin/api/v1/operations/days/{fortuneDate}:
    get:
      operationId: getAdminOperationsDay
      tags: [Admin operations]
      summary: 查看某个命理日的真实用户端投影和订正能力
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/FortuneDate"
      responses:
        "200":
          description: 返回日级摘要、共享公开展示投影与并发保护信息
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDayDetail"
        "400":
          $ref: "#/components/responses/InvalidFortuneDate"
        "401":
          $ref: "#/components/responses/Unauthenticated"
...
  /admin/api/v1/day-corrections:
    post:
      operationId: openDayCorrectionWorkingCopy
      tags: [Admin lifecycle]
      summary: 创建或复用某个命理日的可视化订正工作副本
      description: |
        服务端从该日当前排期版本、活跃版本或最新可用批准版本复制工作副本；同一命理日
        同时只保留一个未完成订正。普通维护者不需要理解草稿或内容版本。
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/XCsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenDayCorrectionRequest"
      responses:
        "200":
          description: 返回新建或既有订正工作副本
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            ETag:
              $ref: "#/components/headers/DayCorrectionETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayCorrectionWorkingCopy"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/CsrfValidationFailed"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /admin/api/v1/day-corrections/{correctionId}:
    get:
      operationId: getDayCorrectionWorkingCopy
      tags: [Admin lifecycle]
      summary: 读取订正工作副本与当前修订号
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/CorrectionId"
      responses:
        "200":
          description: 返回订正工作副本
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            ETag:
              $ref: "#/components/headers/DayCorrectionETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayCorrectionWorkingCopy"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
    patch:
      operationId: patchDayCorrectionWorkingCopy
      tags: [Admin lifecycle]
      summary: 使用字段白名单命令订正文案、穿搭说明或指定图片槽
      description: |
        只接受契约列出的语义命令，不接受 JSON Pointer 或任意模块 JSON。fortuneDate、五档颜色、
        五档顺序、五行结果、业务时间和时辰等算法字段在服务端硬拒绝。If-Match 必须原样回传
        读取工作副本时收到的 opaque strong day-correction ETag；客户端不得自行拼接。成功后返回同时绑定 correction 与 draft revision 的新 ETag。
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/CorrectionId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/XCsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DayCorrectionCommand"
      responses:
        "200":
          description: 字段已更新并增加草稿修订号
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            ETag:
              $ref: "#/components/headers/DayCorrectionETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayCorrectionPatchResult"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/CsrfValidationFailed"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
        "409":
          $ref: "#/components/responses/InvalidStateTransition"
        "412":
          $ref: "#/components/responses/RevisionMismatch"
        "422":
          $ref: "#/components/responses/ImageSetInvalid"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
  /admin/api/v1/day-corrections/{correctionId}/apply:
    post:
      operationId: applyDayCorrection
      tags: [Admin lifecycle]
      summary: 保存订正并按当前公开交付日立即替换或按 18:00 公开边界排期
      description: |
        服务端从同一时刻解析 RequestContext 与 PublicContentContext。目标等于 servedFortuneDate 时
        创建批准版本并立即替换；目标晚于 servedFortuneDate 时使用 PublicContentWindowResolver
        给出的 effectiveFrom（目标日期前一日北京时间 18:00）排期；过去日期拒绝。23:00 命理
        fortuneDate 只服务算法和时辰，不能用于选择立即替换或排期模式。
        If-Match 必须原样回传最后一次读取或订正工作副本时收到的 opaque strong day-correction ETag；客户端不得自行拼接。
        外部 Idempotency-Key 派生稳定且彼此隔离的 submit 与 publish/schedule 子键。若提交成功但
        发布或排期暂时失败，批准版本与订正进度会保留，使用相同外部键重试只恢复发布步骤，
        不创建第二个内容版本。发布与排期继续使用生命周期 revision、预期活跃版本、事务审计、
        排期栅栏和公共缓存清理意图。
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/CorrectionId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/XCsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyDayCorrectionRequest"
      responses:
        "200":
          description: 已立即替换或已安排在固定 18:00 公开内容边界生效
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            ETag:
              $ref: "#/components/headers/DayCorrectionETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayCorrectionApplyResult"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/CsrfValidationFailed"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/RevisionMismatch"
        "422":
          description: 图片集合不安全（IMAGE_SET_INVALID）或发布前检查未通过（PUBLISH_PRECHECK_FAILED）
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                invalidImageSet:
                  value:
                    error:
                      code: IMAGE_SET_INVALID
                      message: 当前图片集合不能安全发布。
                      retryable: false
                      requestId: req-correction-422-image
                      details: {}
                publishPrecheckFailed:
                  value:
                    error:
                      code: PUBLISH_PRECHECK_FAILED
                      message: 发布前检查未通过。
                      retryable: false
                      requestId: req-correction-422-preflight
                      details:
                        preflightChecks: []
        "428":
          $ref: "#/components/responses/PreconditionRequired"
        "503":
          $ref: "#/components/responses/AdminServiceUnavailable"
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: five_admin_session
      description: |
        Issue #31 已确认的同源安全会话。生产 Cookie 固定为 `HttpOnly`、`Secure`、
        `SameSite=Strict`、`Path=/admin` 且无 `Domain`；闲置 30 分钟、绝对 12 小时失效。
        所有后台写操作还必须校验可信同源 `Origin` 与会话绑定的 `X-CSRF-Token`。
  parameters:
    FortuneDate:
      name: fortuneDate
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/FortuneDate"
...
    CorrectionId:
      name: correctionId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/OpaqueId"
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: |
        草稿写操作使用响应中的草稿 ETag；生命周期写操作使用内容版本 ETag；
        全局紧急控制使用公开内容开关 ETag。调用方必须原样回传，不能自行拼接。
      schema:
        type: string
        minLength: 3
        maxLength: 256
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 同一业务意图的网络重试必须复用同一个高熵不透明值。
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: '^[-A-Za-z0-9_:.]{16,128}$'
    XCsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: |
        当前后台会话绑定的随机 CSRF 令牌。后台写操作还必须由服务端校验可信同源 `Origin`；
        令牌只能保存在运行中页面内存，不能放入 localStorage、URL 或日志。
      schema:
        type: string
        minLength: 32
        maxLength: 256
  headers:
...
    XRequestId:
      description: 当前外部请求的追踪编号；缓存命中时也必须按请求生成。
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    DayCorrectionETag:
      description: |
        同时绑定 correctionRevision 与 draftRevision 的不透明强 ETag。客户端只能原样保存和回传，
        不得从响应体字段自行拼接。
      schema:
        type: string
        example: '"dc:MTI6Nw"'
...
    NoStoreCacheControl:
      description: 认证、恢复、安全记录和紧急控制响应不得被浏览器或中间层保存。
      schema:
        type: string
        const: no-store
...
  schemas:
    FortuneDate:
      type: string
      format: date
      description: Five 命理日；北京时间 23:00 进入下一命理日。
      example: "2026-07-15"
    ZonedDateTime:
      type: string
      format: date-time
      example: "2026-07-14T23:30:00+08:00"
...
    OpaqueId:
      type: string
      minLength: 1
      maxLength: 128
      description: 不透明标识；调用方不得解析。
...
    RequestContext:
      type: object
      additionalProperties: false
      required:
        - responseGeneratedAt
        - civilDate
        - fortuneDate
        - shichen
        - timezone
        - dayBoundary
        - crossedDayBoundary
      properties:
        responseGeneratedAt:
          $ref: "#/components/schemas/ZonedDateTime"
        civilDate:
          type: string
          format: date
        fortuneDate:
          $ref: "#/components/schemas/FortuneDate"
        shichen:
          type: string
          enum: [子, 丑, 寅, 卯, 辰, 巳, 午, 未, 申, 酉, 戌, 亥]
        timezone:
          type: string
          const: Asia/Shanghai
        dayBoundary:
          type: string
          const: "23:00"
        crossedDayBoundary:
          type: boolean
    PublicContentContext:
      type: object
      additionalProperties: false
      required: [servedFortuneDate, switchBoundary, advancedFromCivilDate]
      description: |
        匿名用户和普通后台当前实际使用的公开内容日期。它只决定内容选择，
        不改变 23:00 命理日、时辰或日历算法。
      properties:
        servedFortuneDate:
          $ref: "#/components/schemas/FortuneDate"
        switchBoundary:
          type: string
          const: "18:00"
        advancedFromCivilDate:
          type: boolean
...
    AdminDayConcurrency:
      type: object
      additionalProperties: false
      required: [activeContentVersion, lifecycleRevision, scheduleSlotRevision]
      properties:
        activeContentVersion:
          $ref: "#/components/schemas/NullableContentVersion"
        lifecycleRevision:
          type: integer
          minimum: 0
        scheduleSlotRevision:
          type: integer
          minimum: 0
    AdminDayDetail:
      type: object
      additionalProperties: false
      required:
        - requestContext
        - publicContentContext
        - nextOperationalBoundaryAt
        - summary
        - previewSource
        - preview
        - previewRequestContext
        - previewPublicContentContext
        - editableSelectionKeys
        - readonlySelectionKeys
        - concurrency
      properties:
        requestContext:
          $ref: "#/components/schemas/RequestContext"
        publicContentContext:
          $ref: "#/components/schemas/PublicContentContext"
        nextOperationalBoundaryAt:
          $ref: "#/components/schemas/ZonedDateTime"
          description: 下一次运营刷新边界，只能是下一公开内容日的 prepareBy 或 18:00 公开生效时刻，不表示 23:00 命理换日或普通时辰边界。
        summary:
          $ref: "#/components/schemas/AdminDaySummary"
        previewSource:
          type: string
          enum: [published, scheduled, approved, draft, none]
        preview:
          oneOf:
            - $ref: "#/components/schemas/DailyContent"
            - type: "null"
        previewRequestContext:
          $ref: "#/components/schemas/RequestContext"
        previewPublicContentContext:
          $ref: "#/components/schemas/PublicContentContext"
          description: 服务端基于 previewRequestContext 解析的预览公开交付上下文，客户端不得自行计算 servedFortuneDate 或 advancedFromCivilDate。
        editableSelectionKeys:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 128
        readonlySelectionKeys:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 128
        concurrency:
          $ref: "#/components/schemas/AdminDayConcurrency"
...
    DraftRevision:
      type: integer
      minimum: 1
...
    DayCorrectionRevision:
      type: integer
      minimum: 1
    DayCorrectionStatus:
      type: string
      enum: [open, applying, submitted, applied, abandoned]
    DayCorrectionApplyMode:
      type: string
      enum: [immediate, scheduled]
    OpenDayCorrectionRequest:
      type: object
      additionalProperties: false
      required: [fortuneDate]
      properties:
        fortuneDate:
          $ref: "#/components/schemas/FortuneDate"
    DayCorrectionWorkingCopy:
      type: object
      additionalProperties: false
      required:
        - correctionId
        - fortuneDate
        - draftId
        - correctionRevision
        - draftRevision
        - status
        - applyMode
        - sourceContentVersion
        - baselineActiveContentVersion
        - submittedContentVersion
        - createdAt
        - updatedAt
        - modules
      properties:
        correctionId:
          $ref: "#/components/schemas/OpaqueId"
        fortuneDate:
          $ref: "#/components/schemas/FortuneDate"
        draftId:
          $ref: "#/components/schemas/OpaqueId"
        correctionRevision:
          $ref: "#/components/schemas/DayCorrectionRevision"
        draftRevision:
          $ref: "#/components/schemas/DraftRevision"
        status:
          $ref: "#/components/schemas/DayCorrectionStatus"
        applyMode:
          oneOf:
            - $ref: "#/components/schemas/DayCorrectionApplyMode"
            - type: "null"
        sourceContentVersion:
          $ref: "#/components/schemas/NullableContentVersion"
        baselineActiveContentVersion:
          $ref: "#/components/schemas/NullableContentVersion"
        submittedContentVersion:
          $ref: "#/components/schemas/NullableContentVersion"
        createdAt:
          $ref: "#/components/schemas/ZonedDateTime"
        updatedAt:
          $ref: "#/components/schemas/ZonedDateTime"
        modules:
          $ref: "#/components/schemas/DraftModules"
...
    DayCorrectionCommand:
      oneOf:
        - $ref: "#/components/schemas/SetTierExplanationCommand"
        - $ref: "#/components/schemas/SetOutfitFormulaDisclaimerCommand"
        - $ref: "#/components/schemas/SetOutfitFormulaTitleCommand"
        - $ref: "#/components/schemas/SetBasisDisclaimerCommand"
        - $ref: "#/components/schemas/SetBalanceSuggestionDescriptionCommand"
        - $ref: "#/components/schemas/SetShareCopyCommand"
        - $ref: "#/components/schemas/ReplaceImageCoverCommand"
      discriminator:
        propertyName: kind
        mapping:
          set_tier_explanation: "#/components/schemas/SetTierExplanationCommand"
          set_outfit_formula_disclaimer: "#/components/schemas/SetOutfitFormulaDisclaimerCommand"
          set_outfit_formula_title: "#/components/schemas/SetOutfitFormulaTitleCommand"
          set_basis_disclaimer: "#/components/schemas/SetBasisDisclaimerCommand"
          set_balance_suggestion_description: "#/components/schemas/SetBalanceSuggestionDescriptionCommand"
          set_share_copy: "#/components/schemas/SetShareCopyCommand"
          replace_image_cover: "#/components/schemas/ReplaceImageCoverCommand"
    DayCorrectionPatchResult:
      type: object
      additionalProperties: false
      required: [correctionId, fortuneDate, correctionRevision, draftRevision, moduleCode]
      properties:
        correctionId:
          $ref: "#/components/schemas/OpaqueId"
        fortuneDate:
          $ref: "#/components/schemas/FortuneDate"
        correctionRevision:
          $ref: "#/components/schemas/DayCorrectionRevision"
        draftRevision:
          $ref: "#/components/schemas/DraftRevision"
        moduleCode:
          $ref: "#/components/schemas/ModuleCode"
    ApplyDayCorrectionRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 2000
          pattern: ".*\\S.*"
    DayCorrectionApplyResult:
      type: object
      additionalProperties: false
      required: [correctionId, correctionRevision, draftRevision, mode, action]
      properties:
        correctionId:
          $ref: "#/components/schemas/OpaqueId"
        correctionRevision:
          $ref: "#/components/schemas/DayCorrectionRevision"
        draftRevision:
          $ref: "#/components/schemas/DraftRevision"
        mode:
          $ref: "#/components/schemas/DayCorrectionApplyMode"
        action:
          $ref: "#/components/schemas/LifecycleActionResult"
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
  responses:
    InvalidArgument:
      description: 请求字段或格式无效
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ImageSetInvalid:
      description: 草稿视觉模块引用的图片集合不再与服务端候选、审核结果或全局下线事实一致（`IMAGE_SET_INVALID`）
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalidAssetReference:
              $ref: "#/components/examples/ImageSetInvalidAssetReferenceError"
            withdrawnAsset:
              $ref: "#/components/examples/ImageSetWithdrawnAssetError"
...
    InvalidFortuneDate:
      description: 命理日格式或允许范围无效
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthenticated:
      description: 后台会话不存在或已失效
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    AdminServiceUnavailable:
      description: 后台凭据库或安全控制服务暂时不可用；调用方应保留当前输入并稍后重试
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Retry-After:
          schema:
            type: integer
            minimum: 1
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    CsrfValidationFailed:
      description: 后台写请求的可信同源 `Origin` 或会话绑定 CSRF 令牌校验失败
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Conflict:
      description: 幂等键、活跃版本或状态发生冲突
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ResourceNotFound:
      description: 后台资源或任务不存在
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InvalidStateTransition:
      description: 当前状态不允许执行该动作
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    RevisionMismatch:
      description: "`If-Match` 不是资源当前修订号"
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        ETag:
          description: 当前资源 ETag，调用方必须刷新后再操作。
          schema:
            type: string
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            stale:
              $ref: "#/components/examples/RevisionMismatchError"
...
    PreconditionRequired:
      description: 需要并发保护的写操作缺少 `If-Match`
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...